Create BaseNextFileReferenceService and refactor
